Fix st_plugin mock type

+++ src/stata_interface/debug.rs
@@ -1,5 +1,5 @@
 pub type ST_retcode = i32;
-pub type ST_plugin = i32;
+pub type ST_plugin = u8;
 use std::ffi::CString;
 use std::os::raw::c_char;
